+++ src/main/kotlin/responses/responses.kt
@@ -18,7 +18,7 @@ data class UserResponse(
     val name: String,
     val age: Int?,
     val height: Int?,
-    val weight: Int?,
+    val weight: Double?,
     val gender: String?,
     val goal: String,
     val experienceLevel: String,
